add debug

+++ verify-ec.sh
@@ -1,8 +1,15 @@
 
+if [ $# == 0 ]; then 
+    echo "Missing gitops repo parameter"
+    exit 
+fi 
+
+echo "Clone Gitops Repo" 
 START_DIR=$(pwd)
 rm -rf gitops-repo
 git clone $1 gitops-repo
 
+echo "Clone Jenkins Shared Library" 
 rm -rf jenkins-lib
 git clone https://github.com/redhat-appstudio/tssc-sample-jenkins jenkins-lib
 cp jenkins-lib/resources/common.sh gitops-repo/rhtap
@@ -11,9 +18,10 @@ cp jenkins-lib/resources/init.sh gitops-repo/rhtap
 cp jenkins-lib/resources/verify-deps-exist gitops-repo/rhtap
 cp jenkins-lib/resources/verify-enterprise-contract.sh gitops-repo/rhtap 
 
+echo "Used the same tasks from Jenkins to validate locally"
 cd gitops-repo
-REQUIRED_ENV="MY_QUAY_USER "
-REQUIRED_BINARY="tree "
+REQUIRED_ENV="MY_QUAY_USER COSIGN_SECRET_PASSWORD COSIGN_SECRET_KEY COSIGN_PUBLIC_KEY"
+REQUIRED_BINARY="tree cosign "
 rhtap/verify-deps-exist "$REQUIRED_ENV" "$REQUIRED_BINARY" 
 ERR=$?
 echo "Dependency Error $1 = $ERR" 
@@ -47,7 +55,7 @@ function run () {
     printf "\n" 
     if [ $ERR != 0 ]; then
         echo "Fatal Error code for $1 = $ERR" 
-        cleanup
+        #cleanup
         exit 1
     fi
 }
@@ -58,4 +66,4 @@ run  "rhtap/verify-enterprise-contract.sh"
 
 tree ./results 
 
-cleanup
+#cleanup
